+++ packages/cedar/CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- enable passing a custom fetch implementation into queryFeatures()
 ### Changed
 - bump @esri/arcgis-rest-js dependencies to 1.7.1
 - update dev dependencies
+- use jest mocks to inspect queryFeatures options
 
 ## [1.0.0-beta.8]
 ### Added